requirements.txt
# Core async and HTTP
aiohttp>=3.8.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Solana
solana>=0.30.0
solders>=0.20.0
base58>=2.1.0
websockets>=11.0

# Database
asyncpg>=0.28.0

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# Web scraping
beautifulsoup4>=4.12.2
selenium>=4.15.2
lxml>=4.9.3
webdriver-manager>=4.0.1

# Utilities
python-dateutil>=2.8.0
pydantic>=2.0.0
tenacity>=8.2.0
ratelimit>=2.2.1
tqdm>=4.64.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
